+++ tests/Unit/Schema/Directives/ComplexityDirectiveTest.php
@@ -26,7 +26,6 @@ class ComplexityDirectiveTest extends TestCase
             ->getComplexityFn();
 
         $this->assertSame(100, $complexityFn(10, ['first' => 10]));
-        $this->assertSame(100, $complexityFn(10, [config('lighthouse.pagination_amount_argument') => 10]));
     }
 
     public function testCanSetCustomComplexityResolver(): void