+++ tests/Unit/Schema/Source/SchemaSourceProviderTest.php
@@ -44,13 +44,4 @@ class SchemaSourceProviderTest extends TestCase
 
         $currentDir->deleteDir('schema');
     }
-
-    public function testCanSetRootPath(): void
-    {
-        $this->filesystem->put('foo', 'bar');
-
-        app(SchemaSourceProvider::class)->setRootPath(__DIR__.'/schema/foo');
-
-        $this->assertSame('bar'.PHP_EOL, app(SchemaSourceProvider::class)->getSchemaString());
-    }
 }